+++ src/main/scala/com/rayrobdod/json/union/CborValue.scala
@@ -28,6 +28,7 @@ package com.rayrobdod.json.union
 
 import java.math.MathContext.UNLIMITED
 import scala.language.implicitConversions
+import com.rayrobdod.json.union.Failures.UnsuccessfulTypeCoersion
 
 /**
  * A union type representing primitive types in Cbor objects
@@ -57,41 +58,41 @@ sealed trait CborValue {
 	/**
 	 * Executes and returns `fs(this.s)` if this is a CborValueString, else return a Left with an error message
 	 */
-	final def stringToEither[A](fs:String => Either[(String, Int),A]):Either[(String, Int),A] = {
-		val unexpected = new ReturnLeft("Expected string")
+	final def stringToEither[E >: UnsuccessfulTypeCoersion, A](fs:String => Either[E, A]):Either[E, A] = {
+		val unexpected = new ReturnLeft("String")
 		this.fold(fs, unexpected, unexpected, unexpected, unexpected)
 	}
 	
 	/**
 	 * Executes and returns `fs(this.s)` if this is a CborValueByteStr, else return a Left with an error message
 	 */
-	final def byteArrayToEither[A](fs:Array[Byte] => Either[(String, Int),A]):Either[(String, Int),A] = {
-		val unexpected = new ReturnLeft("Expected string")
+	final def byteArrayToEither[E >: UnsuccessfulTypeCoersion, A](fs:Array[Byte] => Either[E, A]):Either[E, A] = {
+		val unexpected = new ReturnLeft("Byte String")
 		this.fold(unexpected, fs, unexpected, unexpected, unexpected)
 	}
 	
 	/**
 	 * Executes and returns `fi(this.i)` if this is a CborValueNumber which holds an number convertible to integer, else return a Left with an error message.
 	 */
-	final def integerToEither[A](fi:Int => Either[(String, Int),A]):Either[(String, Int),A] = {
-		val number = {n:Rational => n.tryToInt.fold[Either[(String, Int), A]](Left(("Expected integral number", 0))){fi}}
-		val unexpected = new ReturnLeft("Expected integral number")
+	final def integerToEither[E >: UnsuccessfulTypeCoersion, A](fi:Int => Either[E, A]):Either[E, A] = {
+		val unexpected = new ReturnLeft("Int")
+		val number = {n:Rational => n.tryToInt.fold[Either[E, A]](unexpected(n)){fi}}
 		this.fold(unexpected, unexpected, number, unexpected, unexpected)
 	}
 	
 	/**
 	 * Executes and returns `fn(this.i)` if this is a CborValueNumber, else return a Left with an error message.
 	 */
-	final def numberToEither[A](fn:Rational => Either[(String, Int),A]):Either[(String, Int),A] = {
-		val unexpected = new ReturnLeft("Expected number")
+	final def numberToEither[E >: UnsuccessfulTypeCoersion, A](fn:Rational => Either[E, A]):Either[E, A] = {
+		val unexpected = new ReturnLeft("Rational")
 		this.fold(unexpected, unexpected, fn, unexpected, unexpected)
 	}
 	
 	/**
 	 * Executes and returns `fb(this.b)` if this is a CborValueBoolean, else return a Left with an error message
 	 */
-	final def booleanToEither[A](fb:Boolean => Either[(String, Int),A]):Either[(String, Int),A] = {
-		val unexpected = new ReturnLeft("Expected boolean")
+	final def booleanToEither[E >: UnsuccessfulTypeCoersion, A](fb:Boolean => Either[E,A]):Either[E,A] = {
+		val unexpected = new ReturnLeft("Boolean")
 		this.fold(unexpected, unexpected, unexpected, fb, unexpected)
 	}
 }
@@ -161,9 +162,9 @@ object CborValue {
 		case JsonValue.JsonValueNull => CborValue.CborValueNull
 	}
 	
-	private class ReturnLeft(msg:String) extends Function1[Any, Either[(String, Int), Nothing]] with Function0[Either[(String, Int), Nothing]] {
-		def apply():Either[(String, Int), Nothing] = Left(msg, 0)
-		def apply(x:Any):Either[(String, Int), Nothing] = Left(msg, 0)
+	private class ReturnLeft(toType:String) extends Function1[Any, Either[UnsuccessfulTypeCoersion, Nothing]] with Function0[Either[UnsuccessfulTypeCoersion, Nothing]] {
+		def apply():Either[UnsuccessfulTypeCoersion, Nothing] = Left(UnsuccessfulTypeCoersion(CborValueNull, "CborValue", toType))
+		def apply(x:Any):Either[UnsuccessfulTypeCoersion, Nothing] = Left(UnsuccessfulTypeCoersion(x, "CborValue", toType))
 	}